fastapi/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-openai
langgraph
pytest
pytest-asyncio
